Refactor: Move duplicated code to proper locations in repo
Moved three commonly duplicated pieces of code to their proper locations:

1. PersistentContextJsonlWriter
   - From: Duplicated in test_finepdfs_local.py, test_rolmocr.py, finepdfs.py
   - To: src/datatrove/pipeline/writers/jsonl.py
   - Purpose: Workaround for InferenceRunner file truncation bug

2. ExtractInferenceText (renamed from PostProcessOCRResults)
   - From: Duplicated in test_finepdfs_local.py, test_rolmocr.py, finepdfs.py
   - To: src/datatrove/pipeline/inference/post_process.py
   - Purpose: Generic post-processor for extracting text from inference_results

3. rolmocr_query_builder
   - From: Duplicated in test_finepdfs_local.py, finepdfs.py
   - To: src/datatrove/pipeline/inference/query_builders/vision.py
   - Purpose: Convert PDF documents to RolmOCR vision requests

All test scripts updated to import from new locations.
test_rolmocr.py keeps local rolmocr_query_builder_with_debug variant.
