+++ examples_local/test_rolmocr.py
@@ -19,60 +19,15 @@ sys.path.insert(0, 'src')
 
 from datatrove.data import Document, Media, MediaType
 from datatrove.pipeline.inference.run_inference import InferenceRunner, InferenceConfig
+from datatrove.pipeline.inference.post_process import ExtractInferenceText
+from datatrove.pipeline.inference.query_builders.vision import rolmocr_query_builder
 from datatrove.pipeline.inference.utils.page_rendering import render_page_to_base64png_pymupdf
-from datatrove.pipeline.writers.jsonl import JsonlWriter
-from datatrove.pipeline.base import PipelineStep
-from datatrove.pipeline.inference.run_inference import InferenceSuccess
+from datatrove.pipeline.writers.jsonl import PersistentContextJsonlWriter
 from datatrove.executor.local import LocalPipelineExecutor
 from typing import Iterable
 
 
-class PostProcessOCRResults(PipelineStep):
-    """Post-process RolmOCR inference results."""
-
-    def run(self, data: Iterable[Document], rank: int = 0, world_size: int = 1):
-        for document in data:
-            # Extract OCR text from inference results and replace PDF bytes
-            document.text = "\n".join([x.text if isinstance(x, InferenceSuccess) else x.error for x in document.metadata["inference_results"]])
-            # Clean up inference_results metadata
-            del document.metadata["inference_results"]
-            yield document
-
-
-class PersistentContextJsonlWriter(JsonlWriter):
-    """JsonlWriter that keeps file context open across multiple run() calls.
-
-    Workaround for framework bug where InferenceRunner calls post_process_steps
-    separately for each document, causing JsonlWriter to close/reopen files
-    between documents, which truncates the output file.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._context_entered = False
-
-    def run(self, data: Iterable[Document], rank: int = 0, world_size: int = 1):
-        # Enter context only once, on first call
-        if not self._context_entered:
-            self.__enter__()
-            self._context_entered = True
-
-        # Write documents without entering/exiting context
-        for document in data:
-            with self.track_time():
-                self.write(document, rank)
-            yield document
-
-    def __del__(self):
-        # Clean up context when object is destroyed
-        if self._context_entered:
-            try:
-                self.__exit__(None, None, None)
-            except:
-                pass
-
-
-def rolmocr_query_builder(runner: InferenceRunner, doc: Document) -> dict:
+def rolmocr_query_builder_with_debug(runner: InferenceRunner, doc: Document) -> dict:
     """Convert PDF document to RolmOCR vision request.
 
     Follows FinePDFs specification:
@@ -226,10 +181,10 @@ def test_rolmocr_integration():
         pipeline=[
             documents,  # Documents as part of the pipeline
             InferenceRunner(
-                query_builder=rolmocr_query_builder,
+                query_builder=rolmocr_query_builder_with_debug,
                 config=config,
                 post_process_steps=[
-                    PostProcessOCRResults(),  # Extract OCR text from inference_results
+                    ExtractInferenceText(),  # Extract OCR text from inference_results
                     PersistentContextJsonlWriter("examples_local/output/rolmocr_results")
                 ]
             ),
